backend/list_models.py
import os
from dotenv import load_dotenv
from google import genai

# Load environment variables
load_dotenv()

client = genai.Client(
    api_key=os.getenv("GEMINI_API_KEY")
)

for model in client.models.list():
    print(model.name)
